requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==4.1.0
deprecation==2.1.0
distro==1.9.0
elevenlabs==0.2.26
executing==2.2.0
fastapi==0.104.1
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.3
mypy_extensions==1.1.0
numpy==1.25.2
openai==1.3.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.13
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.27.0
tqdm==4.67.1
traitlets==5.14.3
twilio==8.10.0
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
